New: traverse() skips "hidden" paths
The traverse() function will now behave just like the traverse()
function in score.js and will skip libraries that start with an
underscore and those residing in a folder starting with an underscore.
Furthermore, to be able to provide *all* libraries like before,
regardless of path, a new keyword-only parameter `include_hidden` was
introduced.